requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
jsonschema>=4.0
pyperclip>=1.9.0